+++ CHANGELOG.md
@@ -5,3 +5,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+### Added
+- Schema object with any value type added to the trigger
+- Tests added to check for schema
+
+### Changed
+- Options now allows any value type
+- Harness updated with additional use cases
+- Tests updated to allow for schema